+++ packages/screenshoter/CHANGELOG.md
@@ -3,7 +3,11 @@
 
 ## Unreleased
 
+
+## 3.3.5 - 2022/2/8
+
 - add "files" field to the package.json to avoid unnecessary files to be published
+- updated to @applitools/utils@1.2.11 (from 1.2.5)
 
 ## 3.3.4 - 2021/12/23